Check for $null or empty string in home directory
Check for $null or empty string in home directory before running
Remove-TervisADUserHomeDirectory

+++ TervisTechnicalServices.psm1
@@ -176,9 +176,9 @@ function Remove-TervisUser {
     if($UserToReceiveComputerIsMac -and -Not $DeleteFilesWithoutMovingThem) {        
         Send-SupervisorOfTerminatedUserSharedEmailInstructions -UserNameOfTerminatedUser $Identity -UserNameOfSupervisor $IdentityOfUserToReceiveData
         Read-Host "Please move the terminated user's files to the recipient's Mac manually using Teamviewer. Once this has been completed, please run `"Remove-TervisADUserHomeDirectory -Identity $Identity -DeleteFilesWithoutMovingThem`". Press `"Enter`" to acknowledge this message"
-    } elseif ($DeleteFilesWithoutMovingThem) {
+    } elseif ($DeleteFilesWithoutMovingThem -and $ADUser.HomeDirectory) {
             Remove-TervisADUserHomeDirectory -Identity $Identity -DeleteFilesWithoutMovingThem
-    } else {
+    } elseif ($ADUser.HomeDirectory) {
         Remove-TervisADUserHomeDirectory -Identity $Identity -IdentityOfUserToReceiveHomeDirectoryFiles $IdentityOfUserToReceiveData
     }
